requirements-to-freeze.txt
# See http://www.kennethreitz.org/essays/a-better-pip-workflow for motivation

alembic
autopep8
Flask
Flask-HTTPAuth
flask-marshmallow
Flask-SQLAlchemy
gunicorn
marshmallow-sqlalchemy
pytest
pytest-cov
pytest-flask
psycopg2
SQLAlchemy-Searchable
sqlalchemy_utils
